fix author regiser exist , also client

+++ server/controllers/admin/author.controller.js
@@ -5,7 +5,6 @@ const { LoginTemplate } = require('../../email/Template');
 const sendAuthormail = new EmailController();
 const bcrypt = require('bcrypt');
 const AddAuthorController = async (req, res) => {
-    const status = false;
     try {
         const thumbnail = req.file;
         if (!thumbnail) {
@@ -15,18 +14,7 @@ const AddAuthorController = async (req, res) => {
             });
         }
         const author_image = thumbnail.filename;
-
         const { author_name, email, bio, subscription_type } = req.body;
-        const exitsUsers = await AuthorModel.checkUserExists(email);
-        if (exitsUsers.length > 0) {
-            const result = {
-                status: false,
-                message: "Email already exists",
-            }
-            return res.status(401).json({
-                result
-            });
-        }
         // generate passwords for author
         const password = generator.generate({
             length: 10,
@@ -35,7 +23,7 @@ const AddAuthorController = async (req, res) => {
             symbols: true,
         })
         let hashed_password = await bcrypt.hash(password, 10);
-        const data = [author_name, email, author_image, bio, hashed_password];
+        const data = { author_name, email, author_image, bio, hashed_password };
         const result = await AuthorModel.addAuthor(data, subscription_type);
         const maildata = {
             to_user: email,
@@ -82,6 +70,9 @@ const DeleteAuthor = async (req, res) => {
     // for single user
     const { authorid } = req.params;
     const { DeleteItems } = req.body;
+    if (!authorid || !DeleteItems) {
+        throw new Error("Deletion id is required")
+    }
     let result = await AuthorModel.deleteAuthor(DeleteItems || authorid);
     return res.status(200).json(
         result
